+++ src/app/Apis/__init__.py
@@ -16,6 +16,7 @@ from app.Controllers.Authenticated.Task.DelayTaskController import api as delay_
 from app.Controllers.Authenticated.Task.DropTaskController import api as drop_task_v1
 from app.Controllers.Authenticated.Task.TaskActivityController import api as task_activity_v1
 from app.Controllers.Authenticated.Task.TaskController import api as task_v1
+from app.Controllers.Authenticated.Task.TaskPrioritiesController import api as task_priorities_v1
 
 
 # swagger monkey patch
@@ -45,6 +46,7 @@ api.add_namespace(delay_task_v1)
 api.add_namespace(drop_task_v1)
 api.add_namespace(task_activity_v1)
 api.add_namespace(task_v1)
+api.add_namespace(task_priorities_v1)
 
 
 api.add_namespace(health_v1)